Implement OneOf in method Create. Update and Delete with relationship Persons.

+++ Pro.Search.Commands/PersonCommands/DeletePersonCommandHandler.cs
@@ -1,17 +1,16 @@
 ﻿using MediatR;
-using Microsoft.EntityFrameworkCore;
 using Pro.Search.Infraestructure.Context;
 using Pro.Search.Infraestructure.Repositories;
-using Pro.Search.PersonDomains.PersonEngine.Entities;
+using Pro.Search.PersonDomains.PersonEngine.OneOf;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using static Pro.Search.PersonDomains.PersonEngine.OneOf.DeleteResponses;
 
 namespace Pro.Search.PersonCommands
 {
-    public class DeletePersonCommandHandler : IRequestHandler<DeletePersonCommand, List<Persons>>
+    public class DeletePersonCommandHandler : IRequestHandler<DeletePersonCommand, DeleteResponses>
     {
         private readonly IContextDB _context;
         private readonly IPersonsRepository repository;
@@ -22,18 +21,21 @@ namespace Pro.Search.PersonCommands
             this.repository = repository;
         }
 
-        public async Task<List<Persons>> Handle(DeletePersonCommand request, CancellationToken cancellationToken)
+        public async Task<DeleteResponses> Handle(DeletePersonCommand request, CancellationToken cancellationToken)
         {
             _ = request ?? throw new ArgumentNullException(nameof(request));
 
             var pessoas = await this.repository.SearchAllPersonToIdPerson(request.Id_Pessoas, cancellationToken).ConfigureAwait(false);
+
+            if (!pessoas.Any()) 
+                return new BadRequest($"Person with Id {request.Id_Pessoas} not exist in database, try delete person with other Id!");
 
             foreach(var item in pessoas)
             {
                 _context.Pessoas.Remove(item);
             }
             await _context.SaveChangesAsync();
-            return pessoas;
+            return new Success(pessoas);
         }
     }
 }
